root - this space intentionally left blank
This empty commit serves as a root for rebases early in a projects life cycle
and prevents various git-based tools from breaking when they try to display data
about the repository.

For more details, see [Always Start With An Empty Commit](https://bit-booster.com/doing-git-wrong/2017/01/02/git-init-empty/)
for the problems this solves and a history of where the idea came from.